Move some types to types.ts
