+++ src/store/index.ts
@@ -4,56 +4,16 @@ import { createDirectStore } from 'direct-vuex';
 import { Simulation } from 'd3-force';
 
 import {
-  Link, Node, Network, SimulationLink,
+  Link, Node, Network, SimulationLink, State, LinkStyleVariables, LoadError, NestedVariables,
 } from '@/types';
 import api from '@/api';
 import { GraphSpec, RowsSpec, TableRow } from 'multinet';
 import {
-  scaleLinear, ScaleLinear, scaleOrdinal, ScaleOrdinal,
+  scaleLinear, scaleOrdinal,
 } from 'd3-scale';
 import { schemeCategory10 } from 'd3-scale-chromatic';
 
 Vue.use(Vuex);
-
-interface LoadError {
-  message: string;
-  buttonText: string;
-  href: string;
-}
-
-interface NestedVariables {
-  bar: string[];
-  glyph: string[];
-}
-
-interface LinkStyleVariables {
-  width: string;
-  color: string;
-}
-
-interface AttributeRanges {
-  [key: string]: {attr: string; min: number; max: number};
-}
-
-export interface State {
-  workspaceName: string | null;
-  networkName: string | null;
-  network: Network | null;
-  selectedNodes: Set<string>;
-  loadError: LoadError;
-  simulation: Simulation<Node, SimulationLink> | null;
-  renderNested: boolean;
-  markerSize: number;
-  fontSize: number;
-  labelVariable: string;
-  colorVariable: string;
-  selectNeighbors: boolean;
-  nestedVariables: NestedVariables;
-  linkVariables: LinkStyleVariables;
-  attributeRanges: AttributeRanges;
-  nodeColorScale: ScaleOrdinal<string, string>;
-  linkWidthScale: ScaleLinear<number, number>;
-}
 
 const {
   store,
